Generate the UXN1330.ucf file.
Allows easier customizing the if clock frequency.

Reasoning to customize the frequency instead of using the default
100 mhz clock but then just running ifclock slower is because
the dram pll in infrastructure.v cannot synthesize a 300 mhz
frequency with all multiliers/dividers.  Usually this isn't a problem
because many work fine.  However, during synthesis, the clock
frequency period of the pll apparently is ignored and instead the
timing constraint is used to guess the incoming clock frequency.
This causes some multipliers to be too high and synthesis fails.